requirements.txt
streamlit==1.38.0
Pillow==10.4.0
streamlit-drawable-canvas==0.9.3

